+++ src/__tests__/reducer.spec.js
@@ -17,7 +17,13 @@ const FIX_TARGET = "TARGET",
       FIX_COMMAND1 = "COMMAND1",
       FIX_COMMAND2 = "COMMAND2",
       FIX_COMMAND3 = "COMMAND3",
-      FIX_COMMAND4 = "COMMAND4";
+      FIX_COMMAND4 = "COMMAND4",
+      FIX_DIR1 = "DIR1",
+      FIX_NAME1 = "NAME1",
+      FIX_DIR2 = "DIR2",
+      FIX_NAME2 = "NAME2",
+      FIX_DIR3 = "DIR3",
+      FIX_NAME3 = "NAME3";
 
 function addTarget( state, target, selector ) {
   return reducer( state, {
@@ -78,6 +84,20 @@ function addCommand( state, method, groupId, testId ) {
     });
 }
 
+function addSettingsProject( state, dir, name ) {
+  return reducer( state, {
+      type: actions.addSettingsProject,
+      payload: { dir, name }
+    });
+}
+
+function removeSettingsProject( state, dir ) {
+  return reducer( state, {
+      type: actions.removeSettingsProject,
+      payload: dir
+    });
+}
+
 function insertAdjacentCommand( state, method, groupId, testId, position ) {
   return reducer( state, {
       type: actions.insertAdjacentCommand,
@@ -100,6 +120,62 @@ function shiftCommands( res ) {
 
 
 describe( "Reducers", () => {
+
+   describe( "*Settings", () => {
+
+    it( "adds a new project", () => {
+      let state = addSettingsProject( DEFAULT_STATE, FIX_DIR1, FIX_NAME1 );
+      const projects = state.settings.projects;
+      expect( projects[ FIX_DIR1 ] ).toBe( FIX_NAME1 );
+    });
+
+    it( "adds multiple projects and they get sorted", () => {
+      let state = addSettingsProject( DEFAULT_STATE, FIX_DIR1, FIX_NAME1 );
+      state = addSettingsProject( state, FIX_DIR3, FIX_NAME3 );
+      state = addSettingsProject( state, FIX_DIR2, FIX_NAME2 );
+      const projects = Object.entries( state.settings.projects );
+      expect( projects.length ).toBe( 3 );
+      expect( projects[ 0 ][ 0 ] ).toBe( FIX_DIR1 );
+      expect( projects[ 0 ][ 1 ] ).toBe( FIX_NAME1 );
+      expect( projects[ 1 ][ 0 ] ).toBe( FIX_DIR2 );
+      expect( projects[ 1 ][ 1 ] ).toBe( FIX_NAME2 );
+      expect( projects[ 2 ][ 0 ] ).toBe( FIX_DIR3 );
+      expect( projects[ 2 ][ 1 ] ).toBe( FIX_NAME3 );
+    });
+
+    it( "removes projects", () => {
+      let state = addSettingsProject( DEFAULT_STATE, FIX_DIR1, FIX_NAME1 );
+      state = addSettingsProject( state, FIX_DIR3, FIX_NAME3 );
+      state = addSettingsProject( state, FIX_DIR2, FIX_NAME2 );
+      state = removeSettingsProject( state, FIX_DIR1 );
+      const projects = Object.entries( state.settings.projects );
+      expect( projects.length ).toBe( 2 );
+      expect( projects[ 0 ][ 0 ] ).toBe( FIX_DIR2 );
+      expect( projects[ 0 ][ 1 ] ).toBe( FIX_NAME2 );
+      expect( projects[ 1 ][ 0 ] ).toBe( FIX_DIR3 );
+      expect( projects[ 1 ][ 1 ] ).toBe( FIX_NAME3 );
+    });
+
+    it( "overrides project with the same dir", () => {
+      let state = addSettingsProject( DEFAULT_STATE, FIX_DIR1, FIX_NAME1 );
+      state = addSettingsProject( state, FIX_DIR3, FIX_NAME3 );
+      state = addSettingsProject( state, FIX_DIR3, FIX_NAME2 );
+      const projects = Object.entries( state.settings.projects );
+      expect( projects.length ).toBe( 2 );
+      expect( projects[ 1 ][ 1 ] ).toBe( FIX_NAME2 );
+    });
+
+    it( "may have mutltiple projects with same names", () => {
+      let state = addSettingsProject( DEFAULT_STATE, FIX_DIR1, FIX_NAME1 );
+      state = addSettingsProject( state, FIX_DIR3, FIX_NAME1 );
+      state = addSettingsProject( state, FIX_DIR2, FIX_NAME1 );
+      const projects = Object.entries( state.settings.projects );
+      expect( projects.length ).toBe( 3 );
+      expect( projects[ 0 ][ 1 ] ).toBe( FIX_NAME1 );
+    });
+
+  });
+
 
   describe( "*Target", () => {